requirements.txt
google-cloud-aiplatform>=1.36.0
google-adk>=1.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock==3.12.0 
black>=22.0.0
flake8>=6.0.0
mypy>=1.0.0
sendgrid>=6.10.0
python-dotenv>=1.0.0 
google-cloud-firestore>=2.13.1